requirements.txt
PyQt6>=6.0
PyQt6-Qt6>=6.0
PyQt6-sip>=13.0
pyttsx3>=2.90
python-dotenv>=0.15.0
google-generativeai
pydub>=0.25.1
